all the stuff i tested
Interactable - for objects you can interact with. maybe change this to an interface and have another objects/obstacles abstract class which can be used as a parent for a seesaw class and other stuff. maybe just the interface or just the abstract.

SpriteAnimation extends transition class - thats the gist

Gravity is incomplete

Game is essentially the main